+++ internal/models/webhook_events.go
@@ -6,6 +6,7 @@ const (
 	WebhookEventPaymentReceived                  WebhookEvent = "PaymentReceived"
 	WebhookEventPaymentNotConfirmed              WebhookEvent = "PaymentNotConfirmed"
 	WebhookEventWithdrawalFromProcessingReceived WebhookEvent = "WithdrawalFromProcessingReceived"
+	WebhookEventPaymentAMLBlocked                WebhookEvent = "PaymentAMLBlocked"
 )
 
 func (s WebhookEvent) String() string {